refactor(records): add RecordsStore interface
Add RecordsStore interface and refactor records.ts into a postgres
implementation of RecordsStore.
No changes of behavior, only a refactor to be able to have multiple
instance of records store.
