+++ packages/persist/lib/app.ts
@@ -2,12 +2,11 @@ import './tracer.js';
 
 import db from '@nangohq/database';
 import { destroy as destroyLogs } from '@nangohq/logs';
-import { destroy as destroyRecords } from '@nangohq/records';
+import { records } from '@nangohq/records';
 import { getLogger, initSentry, once, report } from '@nangohq/utils';
 
 import { autoDeletingDaemon } from './daemons/autodeleting.daemon.js';
 import { autoPruningDaemon } from './daemons/autopruning.daemon.js';
-import { seenPartitionDaemon } from './daemons/seenPartition.daemon.js';
 import { envs } from './env.js';
 import { pubsub } from './pubsub.js';
 import { server } from './server.js';
@@ -33,7 +32,10 @@ initSentry({ dsn: envs.SENTRY_DSN, applicationName: envs.NANGO_DB_APPLICATION_NA
 let api: Server;
 const autoPruning = autoPruningDaemon();
 const autoDeleting = autoDeletingDaemon();
-const seenPartition = seenPartitionDaemon();
+records.startDaemon({
+    tickIntervalMs: envs.PERSIST_SEEN_PARTITION_INTERVAL_MS,
+    maxAgeMs: envs.PERSIST_SEEN_PARTITION_MAX_AGE_MS
+});
 
 try {
     const pubsubConnect = await pubsub.connect();
@@ -57,11 +59,10 @@ const close = once(() => {
     api.close(async () => {
         await autoPruning.abort();
         await autoDeleting.abort();
-        await seenPartition.abort();
         await destroyLogs();
         await db.knex.destroy();
         await db.readOnly.destroy();
-        await destroyRecords();
+        await records.close();
         await pubsub.disconnect();
 
         logger.close();
