--- packages/records/lib/db/client.ts
@@ -1,6 +1,0 @@
-import knex from 'knex';
-
-import { config, configRead } from './config.js';
-
-export const db = knex(config);
-export const dbRead = configRead ? knex(configRead) : db;
