+++ packages/records/lib/index.ts
@@ -1,16 +1,8 @@
-import { db, dbRead } from './db/client.js';
-import { configRead } from './db/config.js';
+import { PostgresStore } from './stores/postgres/postgres.js';
 
-export * from './db/migrate.js';
-export * as records from './models/records.js';
 export * as format from './helpers/format.js';
 export { Cursor } from './cursor.js';
 export type * from './types.js';
-export { clearDb as clearDbTestsOnly } from './db/test.helpers.js';
+export { clearDb as clearDbTestsOnly } from './stores/postgres/tests/helpers.js';
 
-export async function destroy() {
-    await db.destroy();
-    if (configRead) {
-        await dbRead.destroy();
-    }
-}
+export const records = new PostgresStore();
