+++ packages/records/lib/stores/postgres/config.ts
@@ -1,4 +1,4 @@
-import { envs } from '../env.js';
+import { envs } from '../../env.js';
 
 import type { Knex } from 'knex';
