+++ packages/records/lib/stores/postgres/tests/helpers.ts
@@ -1,5 +1,8 @@
-import { db } from './client.js';
-import { schema } from './config.js';
+import knex from 'knex';
+
+import { config, schema } from '../config.js';
+
+const db = knex(config);
 
 // WARNING: to use only in tests
 export async function clearDb(): Promise<void> {
