+++ packages/server/lib/controllers/records/patchPruneRecords.integration.test.ts
@@ -1,6 +1,6 @@
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
 
-import { format, migrate as migrateRecords, records } from '@nangohq/records';
+import { format, records } from '@nangohq/records';
 import { seeders } from '@nangohq/shared';
 
 import { isError, isSuccess, runServer, shouldBeProtected } from '../../utils/tests.js';
@@ -10,7 +10,7 @@ let api: Awaited<ReturnType<typeof runServer>>;
 describe(`PATCH ${route}`, () => {
     beforeAll(async () => {
         api = await runServer();
-        await migrateRecords();
+        await records.migrate();
     });
     afterAll(() => {
         api.server.close();
