+++ packages/server/lib/migrate.ts
@@ -1,7 +1,7 @@
 import { KnexDatabase } from '@nangohq/database';
 import { migrate as migrateKeystore } from '@nangohq/keystore';
 import { start as migrateLogs } from '@nangohq/logs';
-import { migrate as migrateRecords } from '@nangohq/records';
+import { records } from '@nangohq/records';
 
 import migrate from './utils/migrate.js';
 
@@ -9,6 +9,6 @@ const db = new KnexDatabase({ timeoutMs: 0 }); // Disable timeout for migrations
 await migrate(db);
 await migrateKeystore(db.knex);
 await migrateLogs();
-await migrateRecords();
+await records.migrate();
 
 process.exit(0);
